requirements.txt
APScheduler
python-dotenv
openai
